Make SharedPreferences work with TaskManager class. Add/update stuff in gradle.
Used Gson and Json to make the SharedPreferences save everything correctly. On the first load, the Task Manager will be null so an extra method was created to prevent crashing. Subsequent loads will not be null so they won't have that problem. Updated gradle and added the gson implementation to the gradle.